The autocomplete box wasn't displaying properly for many layouts, e.g., if the input box was in a table with any <td> to its left. You can see an example of this glitch at:
    http://jsfiddle.net/taherh/NDjdz/

The problem was a missing "'collision': 'none'" entry in
the positions attribute of the options.  However, the root issue is that
$.ui.autocomplete.option was being replaced rather than extended
by $.ui.triggeredAutocomplete.

This commit keeps the parent option attribute, and extends it with
the necessary new options for triggeredAutocomplete, and in particular,
keeps $.ui.autocomplete's default for the 'positions' option.

You can see the working autocomplete (represented by this commit) at:

    http://jsfiddle.net/taherh/MbU2T/